+++ back/src/forms/resolvers/Mutation.ts
@@ -9,7 +9,9 @@ import markAsSealed from "./mutations/markAsSealed";
 import markAsSent from "./mutations/markAsSent";
 import signedByTransporter from "./mutations/signedByTransporter";
 import markAsReceived from "./mutations/markAsReceived";
+import markAsAccepted from "./mutations/markAsAccepted";
 import markAsTempStored from "./mutations/markAsTempStored";
+import markAsTempStorerAccepted from "./mutations/markAsTempStorerAccepted";
 import markAsProcessed from "./mutations/markAsProcessed";
 import markAsResealed from "./mutations/markAsResealed";
 import markAsResent from "./mutations/markAsResent";
@@ -30,7 +32,9 @@ const Mutation: MutationResolvers = {
   markAsSent,
   signedByTransporter,
   markAsReceived,
+  markAsAccepted,
   markAsTempStored,
+  markAsTempStorerAccepted,
   markAsResealed,
   markAsResent,
   markAsProcessed,