+++ front/src/dashboard/constants.tsx
@@ -3,13 +3,15 @@ export const statusLabels: { [key: string]: string } = {
   DRAFT: "Brouillon",
   SEALED: "En attente de collecte par le transporteur",
   SENT: "En attente de réception",
-  RECEIVED: "Reçu, en attente de traitement",
+  RECEIVED: "Reçu, en attente d'acceptation ou de refus'",
+  ACCEPTED: "Accepté, en attente de traitement",
   PROCESSED: "Traité",
   AWAITING_GROUP: "Traité, en attente de regroupement",
   GROUPED: "Annexé à un bordereau de regroupement",
   NO_TRACEABILITY: "Regroupé, avec autorisation de perte de traçabilité",
   REFUSED: "Refusé",
-  TEMP_STORED: "Entreposé temporairement ou en reconditionnement",
+  TEMP_STORED: "Arrivé à l'entreposage provisoire, en attente d'acceptation",
+  TEMP_STORER_ACCEPTED: "Entreposé temporairement ou en reconditionnement",
   RESENT: "En attente de réception pour traitement",
   RESEALED:
     "En attente de collecte par le transporteur après entreposage provisoire",
@@ -27,6 +29,8 @@ export const transportModeLabels: { [key: string]: string } = {
 export const statusesWithDynamicActions = [
   FormStatus.Sent,
   FormStatus.Received,
+  FormStatus.Accepted,
   FormStatus.TempStored,
+  FormStatus.TempStorerAccepted,
   FormStatus.Resent,
 ];